Invoice functionality added.

+++ controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require('../util/file');
+
 const Product = require("../models/product");
 const { validationResult } = require('express-validator');
 
@@ -118,6 +120,7 @@ exports.postEditProduct = (req, res, next) => {
         product.title = updatedTitle;
         product.price = updatedPrice;
         if (image) {
+            fileHelper.deleteFile(product.imageUrl);
             product.imageUrl = image.path;
         }
         product.description = updatedDescription;
@@ -149,7 +152,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found!'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id })
+        })
         .then(() => {
             console.log('Destroid PRODUCT!');
             res.redirect('/admin/products');
